Got the model working, but a bit different to what we had discussed.
Program2.cs is the 3-relationship example, with three people 1,2,3.
Program3.cs is a 5-relationship example, with four people.
Program4.cs is a 7-relationship example, with five people.

All hard-coded, but should be possible to automate, although it needs a bit
of planning.

I haven't implemented overlapping groups with multiple shared people,
but that should be possible.

Very few comments included, so we should probably have a meeting to
discuss.